Code refactor: change login process and session to myAccount
