+++ src/ClientApp/interfaces/appSettings.tsx
@@ -1,7 +1,6 @@
 import {ImenuText} from './menu';
 import {InewsLanguage} from './news';
 import {IAddCommentText, ICommentText, IErrorHandling, ILoading, IErrorHandlingText} from './common';
-import {IcurrentUser} from './currentUser';
 import { IloginMenu, ILoginResponse, IloginFormHeader } from './login';
 import { IRegistrationText } from './registration';
 import { IMyAccountAllText } from './myAccount';
@@ -15,9 +14,6 @@ export interface IAppSettings {
     commentText?: ICommentText;   
     myAccountText?: IMyAccountAllText; 
     presentationLanguage?: string;
-    isLogged?: boolean;
-    loggedUser?: IcurrentUser;
-    tryLogin?: string;
     fetchData?: IFetchData;
     pageNotFoundText?: IErrorHandlingText;
 }
@@ -29,9 +25,6 @@ export interface IappAction{
 
 export interface IappActions{
     appGetLanguage?(language: string ): Function;
-    makeLogin?( user: string, password: string ): Function;
-    makeLogout?( user: IcurrentUser ): Function;
-    resetLoginStatus?(): Function;
 }
 
 interface Ipayload {
