+++ src/ClientApp/store/actions/appSettings/common.tsx
@@ -1,6 +1,6 @@
-import { startServerCommunication, serverCommunicationError, endServerCommunication } from "./appSettings";
-import { IErrorHandling, IServerResponse } from "../../interfaces/common";
-import { ERROR_GENERIC } from "../../settings";
+import { startServerCommunication, serverCommunicationError, endServerCommunication } from ".";
+import { IErrorHandling, IServerResponse } from "../../../interfaces/common";
+import { ERROR_GENERIC } from "../../../settings";
 
 export function commonServerAction( dispatch: Function, 
                                     serverCall:Function,
