+++ src/ClientApp/store/actions/home/index.tsx
@@ -1,8 +1,8 @@
 import {  CHANGE_PRESENTATION_LANGUAGE, GET_HOME_DATA, RESET_HOME_DATA } from '../../actionTypes';
 import {ptCode} from '../../../settings';
 import { getHomeDataFromServer } from './homeServerCalls';
-import { IhomeDataServer, IHomeAction } from '../../../interfaces/home';
-import { commonServerAction } from '../common';
+import { IHomeAction } from '../../../interfaces/home';
+import { commonServerAction } from '../appSettings/common';
 import { IServerPayload } from '../../../interfaces/common';
 
 export function getHomeData( language: string = ptCode ) : Function {
